Use WeatherParser.java and, Improve Weather.java, DisasterSubcategory.java and PrefecturalForecastRegion.java and WarningState.java
